deletePost kinda works but weird error

+++ forumPlateau_V2/view/forum/listPosts.php
@@ -7,11 +7,16 @@ $posts = $result["data"]['posts'];
 <h1>Liste des posts</h1>
 
 <?php
-foreach ($posts as $post) { ?>
-    <p><?= $post ?> par <?= $post->getUser() ?>, le <?= $post->getCreationDate() ?><a href="index.php?ctrl=forum&action=deletePost&id=<?= $post->getId()?>">Supprimer Message</a></p>
-<?php } ?>
+if ($posts) {
+    foreach ($posts as $post) { ?>
+        <p><?= $post ?> par <?= $post->getUser() ?>, le <?= $post->getCreationDate() ?><a href="index.php?ctrl=forum&action=deletePost&id=<?= $post->getId() ?>">Supprimer Message</a></p>
+    <?php }
+} else { ?>
+    <p>Il n'y a pas de Posts dans ce topic ☺☻</p>
+<?php }
+?>
 
-<form action="index.php?ctrl=forum&action=addPost&id=<?= $topic->getId()?>" method="post">
+<form action="index.php?ctrl=forum&action=addPost&id=<?= $topic->getId() ?>" method="post">
     <input type="text" name="text" placeholder="Message">
     <input type="submit">
 </form>
